app and arrange_urlpatterns method moved to app/utils.py. The path to gardens app has been reorganized in app/app/urls.py, unused commented out paths removed, and the new 'activities' path added. This consolidates the utils functions into one file and starts the structure for the new 'activities' app.
