+++ app/activities/urls.py
@@ -1,0 +1,11 @@
+from app import utils
+
+app_name = 'activities'
+
+urlpatterns = [
+]
+
+htmx_urlpatterns = [
+]
+
+urlpatterns = utils.arrange_urlpatterns(urlpatterns + htmx_urlpatterns)
