+++ app/app/urls.py
@@ -22,12 +22,10 @@ from app.views import IndexView
 
 urlpatterns = [
     path("", IndexView.as_view(), name='home'),
-    path('gardens/', include('gardens.urls'), name='gardens'),
     path('accounts/', include('accounts.urls'), name='accounts'),
     path('admin/', admin.site.urls),
-    # path('login/', account_views.login_view),
-    # path('logout/', account_views.logout_view, name='logout'),
-    # path('register/', account_views.register_view)
+    path('gardens/', include('gardens.urls'), name='gardens'),
+    path('activities/', include('activities.urls'), name='activities')
 ]
 
 if settings.DEBUG:
